feat(purchase): 添加用户钱包功能并实现预约订单支付
- 新增 UserWalletMapper 和 WalletPayForAppointment 接口
- 添加 UserWalletDo 和 AppointmentOrderStatusBo 类
- 实现 payAppointmentOrder 方法，支持用户使用钱包支付预约订单
- 在数据库中添加 user_wallet 表结构

+++ spring-server/api/api-model/src/main/java/com/czy/api/domain/bo/medicine/AppointmentOrderStatusBo.java
@@ -2,6 +2,7 @@ package com.czy.api.domain.bo.medicine;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 /**
@@ -19,7 +20,7 @@ public class AppointmentOrderStatusBo {
     // user订单状态
     private Integer userOrderStatus;
     // 商户的定价金额
-    private Integer merchantPrice;
+    private BigDecimal merchantPrice;
     // 预约的开始时间
     private LocalDateTime beginDate;
 }
